Divided the code into different files
Split the code into different files to make the code easier to read and to create a better structure for it all.
Added some comments to make it easier to understand the code
